backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
deepgram-sdk==2.12.0
google-generativeai==0.3.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
websockets==12.0
requests==2.31.0 